Add other fields to find
